fixed play/pause functionality
I found that playsound does not allow for pausing audio, so I had to replace the playsound library with pygame which had more detailed audio playback functionality like pausing. So I had to make some changes to the ttsManager class to adjust for the new library, then I made adjustments to the pause_progress_bar and resume_progress_bar functions to make sure the pause and play buttons worked correctly with the tts and the progress bar.